Updated code per Carlos's comments
