+++ prime-router/src/main/kotlin/cli/SettingCommands.kt
@@ -128,8 +128,11 @@ abstract class SettingCommand(
         }
     }
 
-    fun delete(environment: Environment, accessToken: String, settingType: SettingType, settingName: String) {
+    fun delete(environment: Environment, accessToken: String, settingType: SettingType, settingName: String): String {
         val path = formPath(environment, Operation.DELETE, settingType, settingName)
+        if (verbose) {
+            echo("DELETE $path")
+        }
         val (_, response, result) = Fuel
             .delete(path)
             .authentication()
