--- prime-router/src/test/json_test_files/DummyUpdate.json
@@ -1,12 +1,0 @@
-{
-    "name": "dummy",
-    "description": "PRIME's UPDATED Dummy Organization for Test CRUD of the Setting API",
-    "jurisdiction":"FEDERAL",
-    "stateCode": null,
-    "countyName": null,
-    "meta": {
-        "version": 0,
-        "createdBy": "local@test.com",
-        "createdAt": "2021-10-16T13:02:22.382422Z"
-    }
-}
